Fix msvc c4267 warning

+++ src/util.cpp
@@ -1272,7 +1272,7 @@ void ShrinkDebugFile()
         // Restart the file with some of the end
         char pch[200000];
         fseek(file, -((long long)sizeof(pch)), SEEK_END);
-        int nBytes = fread(pch, 1, sizeof(pch), file);
+        size_t nBytes = fread(pch, 1, sizeof(pch), file);
         fclose(file);
 
         file = fopen(pathLog.string().c_str(), "w");
